% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_clinical_input}
\alias{check_clinical_input}
\title{Check clinical input}
\usage{
check_clinical_input(clinical_input)
}
\arguments{
\item{clinical_input}{}
}
\value{
The function returns an error of the format of the "samples" dataframe has errors
}
\description{
Check clinical input
}
