% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_genomic_input}
\alias{check_genomic_input}
\title{Check that the input mutation dataframe is valid}
\usage{
check_genomic_input(mutations)
}
\arguments{
\item{mutations}{}
}
\description{
Check that the input mutation dataframe is valid
}
