% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_parameters_input}
\alias{check_parameters_input}
\title{Checks that the parameters dataframe is valid}
\usage{
check_parameters_input(parameters)
}
\arguments{
\item{parameters}{}
}
\description{
Checks that the parameters dataframe is valid
}
