% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_ppc}
\alias{check_ppc}
\title{Posterior Predictive checks}
\usage{
check_ppc(x)
}
\arguments{
\item{x}{TOSCA object}
}
\value{
dataframe with columns: name, the name of the mutation cluster, pass: whether the ppc test was passed, i.e. the real number of mutations falls within 1 sd from the mean of the posterior predictive distribution
}
\description{
Posterior Predictive checks
}
