% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_cols}
\alias{check_required_cols}
\title{Check that for the required columns for each dataframe are present}
\usage{
check_required_cols(df, type)
}
\arguments{
\item{df}{df to check}

\item{type}{type of df}
}
\value{
Check that for the required columns for each dataframe are present
}
\description{
Check that for the required columns for each dataframe are present
}
