% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_valid_karyotype}
\alias{check_valid_karyotype}
\title{Check that the input karytypes in the mutations dataframe are valid}
\usage{
check_valid_karyotype(karyo)
}
\arguments{
\item{karyo}{}
}
\value{
Check that the input karytypes in the mutations dataframe are valid
}
\description{
Check that the input karytypes in the mutations dataframe are valid
}
