% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_date_real}
\alias{convert_date_real}
\title{Converts date into real number}
\usage{
convert_date_real(date, x, ref_year = 1900)
}
\arguments{
\item{ref_year}{}
}
\description{
Converts date into real number
}
