% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_real_date}
\alias{convert_real_date}
\title{Converts real number into date}
\usage{
convert_real_date(x, date = NULL, ref_year = 1900)
}
\arguments{
\item{ref_year}{}
}
\description{
Converts real number into date
}
