% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MAP.R
\name{days_from}
\alias{days_from}
\title{Computes the days of the event of interest from a date of interest}
\usage{
days_from(x, parameter, from, cna_length, quantiles = c(5, 95))
}
\arguments{
\item{x}{TOSCA obj}

\item{parameter}{the inferred time of interest}

\item{from}{the date of interest (string, "YY-mm-dd")}

\item{cna_length}{if the inferred date refers to a CNA, the length of the CNA}

\item{quantiles}{}
}
\value{
the distance in days between an arbitrary date (parameter from) and a chosen inferred time
}
\description{
Computes the days of the event of interest from a date of interest
}
\examples{
data("exampleFit")
days_from(exampleFit, parameter = "t_cna", from = "2010-01-04", cna_length = 4.5e+08, quantiles=c(5, 95))
}
