% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example dataset}
\format{
A data frame with 5 rows and 2 columns.
}
\usage{
data(exampleData)
}
\description{
A description of the example dataset.
}
\keyword{datasets}
