% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData_CNA_dormancy}
\alias{exampleData_CNA_dormancy}
\title{Dataset with a toy example for the CNA model with dormancy}
\format{
A list of dataframes with the input required for running TOSCA
}
\usage{
data(exampleData_CNA_dormancy)
}
\description{
Dataset with a toy example for the CNA model with dormancy
}
\keyword{datasets}
