% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Infer the timing of the events of interest fitting the appropriate TOSCA model}
\usage{
fit(
  x,
  n_iterations = 10000,
  n_chains = 4,
  warm_up = 5000,
  cores = 4,
  model_name = "Driver",
  dormancy = F,
  verbose = F
)
}
\arguments{
\item{x}{Tosca object}

\item{n_iterations}{int, number of iteration of each MCMC chain}

\item{n_chains}{int, number of chains}

\item{warm_up}{A positive integer specifying the number of warmup (aka burnin) iterations per chain}

\item{cores}{Number of cores to use when executing the chains in parallel}

\item{verbose}{}
}
\description{
Infer the timing of the events of interest fitting the appropriate TOSCA model
}
