% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnotics.R
\name{get_cmdstanr_posterior}
\alias{get_cmdstanr_posterior}
\title{Get posterior draws from cmdstanr obj}
\usage{
get_cmdstanr_posterior(x, days_from = NULL, parameters = NULL)
}
\arguments{
\item{x}{TOSCA obj}

\item{days_from}{date from wich to compute the distance from for times}

\item{parameters}{parameters one desires to retrive the posterior for}
}
\value{
posterior draws
}
\description{
Get posterior draws from cmdstanr obj
}
