% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MAP.R
\name{get_index_from_cna_length}
\alias{get_index_from_cna_length}
\title{Get index from cna_length}
\usage{
get_index_from_cna_length(x, len)
}
\arguments{
\item{x}{TOSCA obj}

\item{len}{length of the CNA of interest}
}
\value{
index of the corresponding CNA in the model
}
\description{
Get index from cna_length
}
