% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MAP.R
\name{get_index_from_drug_name}
\alias{get_index_from_drug_name}
\title{Get index from drug name}
\usage{
get_index_from_drug_name(x, drug_name)
}
\arguments{
\item{x}{TOSCA obj}

\item{drug_name}{name of the drug}
}
\value{
index of the corresponding drug in the model
}
\description{
Get index from drug name
}
