% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inference_data.R
\name{get_inference_data}
\alias{get_inference_data}
\title{Get inference data}
\usage{
get_inference_data(x, model, dormancy = F, max_mrca = NA, reg_dormancy = 0)
}
\arguments{
\item{x}{TOSCA obj}

\item{model}{string, either "CNA" or "Driver"}

\item{dormancy}{boolean, if model == "CNA", it is possible to include dormancy}

\item{max_mrca}{date in "YYYY-mm-dd" format, maximum date at which the MRCA could have been born (if < than the date of the second sample)}

\item{reg_dormancy}{boolean (0 or 1), optional regularisation of the growth imposed in case of dormancy}
}
\value{
the list of arguments passed to the stan fit function
}
\description{
Get inference data
}
