% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_vs_posterior.R
\name{get_name_mu_step}
\alias{get_name_mu_step}
\title{Title}
\usage{
get_name_mu_step(x, index)
}
\arguments{
\item{x}{TOSCA object}

\item{index}{index of mu_step parameter}
}
\value{
name of the drug corresponding to the index
}
\description{
Title
}
