% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_prior_distribution_type}
\alias{get_prior_distribution_type}
\title{Return prior distribution type for each parameter}
\usage{
get_prior_distribution_type(par)
}
\arguments{
\item{par}{Name of the parameter of interest.}
}
\value{
Name of the prior distribution used in the model
}
\description{
Return prior distribution type for each parameter
}
