% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialise a TOSCA object}
\usage{
init(
  mutations,
  clinical_records,
  parameters,
  delta_omega = 0.02,
  growth_rate_variance = 10
)
}
\arguments{
\item{mutations}{dataframe containing input mutations}

\item{parameters}{}
}
\description{
Initialise a TOSCA object
}
