% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_ggplot_theme}
\alias{my_ggplot_theme}
\title{Customized ggplot theme}
\usage{
my_ggplot_theme(cex = 1)
}
\arguments{
\item{cex}{}
}
\value{
Customized ggplot theme
}
\description{
Customized ggplot theme
}
