% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expected_N.R
\name{plot_expected_N}
\alias{plot_expected_N}
\title{Posterior distribution of the number of cells collected in the first and second samples.}
\usage{
plot_expected_N(x)
}
\arguments{
\item{x}{TOSCA object}
}
\value{
Posterior distribution of the number of cells collected in the first and second samples, considering the growth rate (omega) and the time of birth of the MRCA (t_mrca) of each sample.
}
\description{
Posterior distribution of the number of cells collected in the first and second samples.
}
