% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnotics.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Plot univariate and multivariate distributions of selected parameters}
\usage{
plot_pairs(x, days_from = NULL, parameters = NULL, nuts_params = "divergent__")
}
\arguments{
\item{x}{TOSCA object}

\item{days_from}{date from which to compute the distance from for times}

\item{parameters}{parameters one desires to retrieve the posterior for}

\item{nuts_params}{annotations from bayes plot ("lp__", "energy__", "stepsize__","n_leapfrog__", "accept_stat__", "treedepth__")}
}
\value{
plot
}
\description{
Plot univariate and multivariate distributions of selected parameters
}
