% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_predictive_checks.R
\name{plot_ppc}
\alias{plot_ppc}
\title{Plot posterior predictive checks}
\usage{
plot_ppc(x)
}
\arguments{
\item{x}{TOSCA object}
}
\value{
plot of the posterior predictive distribution versus the real value of all mutation groups provided in input
}
\description{
Plot posterior predictive checks
}
\examples{
library(TOSCA)
library(dplyr)
library(ggplot2)
data("exampleData_CNA")
mutations = exampleData_CNA$Mutations
parameters = exampleData_CNA$Parameters
samples = exampleData_CNA$Samples
therapies = exampleData_CNA$Therapies

x = init(mutations=mutations, samples=samples, therapies=therapies, parameters=parameters)
fit = TOSCA::fit(x, model_name='CNA', n_iterations = 1000, n_chains = 4, warm_up = 500)
plot_ppc(fit)
}
