% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_predictive_checks.R
\name{plot_ppc_single_mut}
\alias{plot_ppc_single_mut}
\title{Plot the posterior predictive distribution and compares it to the real number of mutations}
\usage{
plot_ppc_single_mut(x, mut1_real, mut2_real, rep_name1, rep_name2)
}
\arguments{
\item{x}{TOSCA object}

\item{mut1_real}{real value of the first mutation type of interest}

\item{mut2_real}{real value of the second mutation type of interest}

\item{rep_name1}{name of the posterior predictive variable corresponding to the first mutation type of interest, as reported in the model}

\item{rep_name2}{name of the posterior predictive variable corresponding to the second mutation type of interest, as reported in the model}
}
\value{
plot
}
\description{
Plot the posterior predictive distribution and compares it to the real number of mutations
}
