% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prior_vs_posterior}
\alias{plot_prior_vs_posterior}
\title{Produce collective plot of prior vs posterior for all inferred parameters}
\usage{
plot_prior_vs_posterior(x)
}
\arguments{
\item{x}{TOSCA obj}
}
\value{
plot
}
\description{
Produce collective plot of prior vs posterior for all inferred parameters
}
