% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prior_vs_posterior_single_parameter}
\alias{plot_prior_vs_posterior_single_parameter}
\title{Prior vs Posterior distribution of inferred parameter}
\usage{
plot_prior_vs_posterior_single_parameter(x, parameter)
}
\arguments{
\item{x}{TOSCA object}

\item{parameter}{parameter name}
}
\value{
ggplot
}
\description{
Prior vs Posterior distribution of inferred parameter
}
