% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timing.R
\name{plot_timing_density}
\alias{plot_timing_density}
\title{Plot clinical timeline + posterior times with density}
\usage{
plot_timing_density(x)
}
\arguments{
\item{x}{}
}
\description{
Plot clinical timeline + posterior times with density
}
