% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TOSCA.R
\name{print.TOSCA}
\alias{print.TOSCA}
\title{Print for class \code{'TOSCA'}.}
\usage{
\method{print}{TOSCA}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'TOSCA'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'TOSCA'}.
}
\examples{
\dontrun{
data('exampleFit', package = 'TOSCA')

print(exampleFit)
}
}
