% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWhatsApp.R
\name{readWhatsApp}
\alias{readWhatsApp}
\title{Read WhatsApp files}
\usage{
readWhatsApp(path, file)
}
\arguments{
\item{path}{Character: string with path where the data files are.
If only \code{path} is given, \code{file} will be determined by searching
for html files with \code{\link[base]{list.files}} and recursion.}

\item{file}{Character: string with names of the HTML files.}
}
\value{
\code{\link{textmeta}} object.
}
\description{
Reads HTML-files from WhatsApp and separates the text and meta data.
}
\author{
Jonas Rieger (<jonas.rieger@tu-dortmund.de>)
}
\keyword{manip}
