% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikinews.R
\name{readWikinews}
\alias{readWikinews}
\title{Read files from Wikinews}
\usage{
readWikinews(
  path = getwd(),
  file = list.files(path = path, pattern = "*.xml$", full.names = FALSE, recursive =
    TRUE)
)
}
\arguments{
\item{path}{Path where the data files are.}

\item{file}{Character string with names of the HTML files.}
}
\value{
textmeta-object
}
\description{
Reads the XML-files from the Wikinews export page \url{https://en.wikinews.org/wiki/Special:Export}.
}
