% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnotics.R
\name{rhats_plot}
\alias{rhats_plot}
\title{Plot Rhats of selected parameters}
\usage{
rhats_plot(x, parameters = NULL)
}
\arguments{
\item{x}{TOSCA object}

\item{parameters}{}
}
\value{
plot
}
\description{
Plot Rhats of selected parameters
}
