% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Sample Texts}
\usage{
sampling(id, corporaID, label, m, randomize = FALSE, exact = FALSE)
}
\arguments{
\item{id}{Character: ID of all texts in the corpus.}

\item{corporaID}{List of Character: Each list contains the IDs belonging to one subcorpus. Each ID has to be in \code{id}.}

\item{label}{Named Logical: Labeling result for already labeled texts. Names must be corresponding \code{id}.}

\item{m}{Integer: Number of new samples.}

\item{randomize}{Logical: If \code{TRUE} calculated split is used as parameter to draw from a multnomial distribution.}

\item{exact}{Logical: If \code{TRUE} exact calculation is used. For the default \code{FALSE} an approximation is used.}
}
\value{
\code{\link{textmeta}} object if \code{object} is specified,
else only the filtered \code{text}. If a \code{\link{textmeta}} object is
returned its meta data are filtered to those texts which appear in the corpus.
}
\description{
Sample texts from different subsets to minimize variance of the recall estimator
}
\examples{
id <- paste0("ID", 1:1000)
corporaID <- list(sample(id, 300), sample(id, 100), sample(id, 700))
label <- sample(as.logical(0:1), 150, replace=TRUE)
names(label) <- c(sample(id, 100), sample(corporaID[[2]], 50))
m <- 100
sampling(id, corporaID, label, m, exact=FALSE)
sampling(id, corporaID, label, m, exact=TRUE)
sampling(id, corporaID, label, m, randomize=TRUE)
}
