% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTextTopics.R
\name{showTextTopics}
\alias{showTextTopics}
\title{Export Topic-Matrices per Text}
\usage{
showTextTopics(ldaresult, documents, id = names(documents),
  vocab = colnames(ldaresult$topics), tnames, file, prefix = "",
  fileEncoding = "UTF-8")
}
\arguments{
\item{ldaresult}{LDA result.}

\item{documents}{A list prepared by \code{\link{LDAprep}}.}

\item{id}{Character vector containing the considered text ids
(default is \code{names(documents)}).}

\item{vocab}{Character vector containing the words in the corpus
(default is \code{colnames(ldaresult$topics)}).}

\item{tnames}{Character vector containing the names for the topics
(default is \code{1:K},
while \code{K} is the number of topics assumed in \code{ldaresult}).}

\item{file}{Character File path for the export. If not specified the function
does not write. If \code{file = ""} the current file path is chosen.}

\item{prefix}{Character Prefix for each file (each text). File names are
build from prefix and the corresponding text id.}

\item{fileEncoding}{Character Declares file encoding. For more information
see \code{\link[utils]{write.csv}}.}
}
\value{
A list of topic-word-matrices of the requested texts.
If file is set, for each text a csv is written.
}
\description{
The function extracts the matrix (matrices) of assignments per topic and word in texts.
}
\examples{
texts <- list(
 A = "Give a Man a Fish, and You Feed Him for a Day.
      Teach a Man To Fish, and You Feed Him for a Lifetime",
 B = "So Long, and Thanks for All the Fish",
 C = "A very able manipulative mathematician, Fisher enjoys a real mastery
      in evaluating complicated multiple integrals.")

corpus <- textmeta(
 meta = data.frame(
  id = c("A", "B", "C", "D"),
  title = c("Fishing", "Don't panic!", "Sir Ronald", "Berlin"),
  date = c("1885-01-02", "1979-03-04", "1951-05-06", "1967-06-02"),
  additionalVariable = 1:4,
  stringsAsFactors = FALSE),
 text = texts)

corpus <- cleanTexts(corpus)
wordlist <- makeWordlist(corpus$text)
ldaPrep <- LDAprep(text=corpus$text, vocab=wordlist$words)

\donttest{LDA <- LDAgen(documents = ldaPrep, K = 3L, vocab = wordlist$words, num.words = 3)}
\donttest{res <- showTextTopics(ldaresult = LDA, documents = ldaPrep)}
}
\keyword{manip}
