% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPN06}
\alias{UPN06}
\title{Input data for patient UPN06}
\format{
A list of dataframes with the input of TOSCA
}
\usage{
data(UPN06)
}
\description{
Input data for patient UPN06
}
\keyword{datasets}
