% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikinews.R
\docType{data}
\name{wikinews}
\alias{wikinews}
\alias{politics}
\alias{economy}
\title{The wikinews dataset}
\format{textmeta object}
\usage{
data(politics)

data(economy)
}
\description{
Two datasets generated from wikinews.
The wikinews data is published under CC2.5 (\url{https://creativecommons.org/licenses/by/2.5/}).
}
\details{
The economy dataset contains 2327 articles between 2004-11-17 and 2018-03-04 that where assigned to the category Economy_and_business.
The politics dataset contains 5000 articles between 2004-11-13 and 2009-11-19 that where assigned to the category Politics_and_conflicts.
}
\references{
\url{https://en.wikinews.org/wiki/Special:Export}
}
\keyword{datasets}
