% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Infer the timing of the events of interest fitting the appropriate TOSCA model}
\usage{
fit(
  x,
  n_iterations = 10000,
  n_chains = 4,
  warm_up = 5000,
  cores = 4,
  model_name = "Driver",
  dormancy = F,
  verbose = F,
  initialisation = F,
  init_fun = NULL
)
}
\arguments{
\item{x}{TOSCA object}

\item{n_iterations}{Number of iteration of each MCMC chain. The default is 10000.}

\item{n_chains}{A positive integer specifying the number of Markov chains. The default is 4.}

\item{warm_up}{A positive integer specifying the number of warmup (aka burnin) iterations per chain. The default is 5000.}

\item{cores}{Number of cores to use when executing the chains in parallel}

\item{model_name}{Name of the model: "CNA" or "Driver"}

\item{dormancy}{Boolean specifying if the model should include dormancy (available only for "CNA" model). The default is F}

\item{verbose}{Boolen specifying if the fit function should output the iterations progression. The default is F.}
}
\value{
TOSCA object
}
\description{
Infer the timing of the events of interest fitting the appropriate TOSCA model
}
