% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ppc}
\alias{plot_ppc}
\title{Plot posterior predictive checks}
\usage{
plot_ppc(x)
}
\arguments{
\item{x}{TOSCA object}
}
\value{
plot of the posterior predictive distribution versus the real value of all mutation groups provided in input
}
\description{
Plot posterior predictive checks
}
