% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prior_vs_posterior}
\alias{plot_prior_vs_posterior}
\title{Produce collective plot of prior vs posterior for all inferred parameters}
\usage{
plot_prior_vs_posterior(x)
}
\arguments{
\item{x}{TOSCA obj}
}
\value{
plot
}
\description{
Produce collective plot of prior vs posterior for all inferred parameters
}
\examples{
library(TOSCA)
library(dplyr)
library(ggplot2)
data("exampleData_CNA")
mutations = exampleData_CNA$Mutations
parameters = exampleData_CNA$Parameters
samples = exampleData_CNA$Samples
therapies = exampleData_CNA$Therapies

x = init(mutations=mutations, samples=samples, therapies=therapies, parameters=parameters)
fit = TOSCA::fit(x, model_name='CNA', n_iterations = 1000, n_chains = 4, warm_up = 500)
plot_prior_vs_posterior(fit)
}
