% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timing.R
\name{plot_timing}
\alias{plot_timing}
\title{Plot clinical timeline + posterior times with histogram}
\usage{
plot_timing(x)
}
\arguments{
\item{x}{}
}
\description{
Plot clinical timeline + posterior times with histogram
}
