% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_timing}
\alias{plot_timing}
\title{Posterior distribution of the inferred times}
\usage{
plot_timing(x)
}
\arguments{
\item{x}{}
}
\value{
Posterior distributio plot of the inferred times, mapped on the clinical history
}
\description{
Posterior distribution of the inferred times
}
