% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timing.R
\name{plot_timing}
\alias{plot_timing}
\title{Posterior distribution of the inferred times}
\usage{
plot_timing(x)
}
\arguments{
\item{x}{TOSCA object}
}
\value{
Posterior distributio plot of the inferred times, mapped on the clinical history
}
\description{
Posterior distribution of the inferred times
}
\examples{
library(TOSCA)
library(dplyr)
library(ggplot2)
data("exampleData_CNA")
mutations = exampleData_CNA$Mutations
parameters = exampleData_CNA$Parameters
samples = exampleData_CNA$Samples
therapies = exampleData_CNA$Therapies

x = init(mutations=mutations, samples=samples, therapies=therapies, parameters=parameters)
fit = TOSCA::fit(x, model_name='CNA', n_iterations = 1000, n_chains = 4, warm_up = 500)
plot_timing(fit)
}
