% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timing.R
\name{plot_timing_MAP}
\alias{plot_timing_MAP}
\title{Plot clinical timeline + posterior times with MAP}
\usage{
plot_timing_MAP(x)
}
\arguments{
\item{x}{}
}
\description{
Plot clinical timeline + posterior times with MAP
}
